refactor(terminal): 拆出快捷命令条/确认框与命令输入/自动补全 hook 及组件
- 新增 components/terminal/：useTerminalQuickCmd（快捷命令条状态/参数历史/确认发送）、useTerminalCommandInput（命令文本/高度自适应/cwd 跟踪/自动补全加载与执行）、TerminalQuickCmdBar（命令条与内嵌搜索）、TerminalInputBar（底部命令输入栏）、TerminalAutocompletePopup（补全弹层）、TerminalQuickCmdConfirm（快捷命令参数确认框与历史弹层）
- Terminal.tsx 2365 -> 1311 行
